Separate courses array into a module
