+++ src/data/courseImages.js
@@ -1,0 +1,71 @@
+import next from "../assets/next.png";
+import react from "../assets/react.png";
+import php from "../assets/php.png";
+import swift from "../assets/swift.png";
+import go from "../assets/go.png";
+import ruby from "../assets/ruby.png";
+import node from "../assets/node.png";
+import python from "../assets/python.png";
+import HTML from "../assets/HTML.png";
+import kotlin from "../assets/kotlin.png";
+import java from "../assets/java.png";
+import Chash from "../assets/Chash.png";
+import CSS from "../assets/CSS.png";
+import CPP from "../assets/CPP.png";
+import JS from "../assets/JS.png";
+import c from "../assets/C.png";
+import groovy from "../assets/groovy.png";
+import assembly from "../assets/assembly.png";
+import dart from "../assets/dart.png";
+import elixir from "../assets/elixir.png";
+import fsharp from "../assets/fsharp.png";
+import haskell from "../assets/haskell.png";
+import lua from "../assets/lua.png";
+import objc from "../assets/objc.png";
+import perl from "../assets/perl.png";
+import prolog from "../assets/prolog.png";
+import R from "../assets/RL.png";
+import rust from "../assets/rust.png";
+import scala from "../assets/scala.png";
+import shell from "../assets/shell.png";
+import sql from "../assets/sql.png";
+import typescript from "../assets/typescript.png";
+import clojure from "../assets/clojure.png";
+
+const courseImages = {
+  next,
+  react,
+  php,
+  swift,
+  go,
+  ruby,
+  node,
+  python,
+  HTML,
+  kotlin,
+  java,
+  Chash,
+  CSS,
+  CPP,
+  JS,
+  c,
+  groovy,
+  assembly,
+  dart,
+  elixir,
+  fsharp,
+  haskell,
+  lua,
+  objc,
+  perl,
+  prolog,
+  R,
+  rust,
+  scala,
+  shell,
+  sql,
+  typescript,
+  clojure,
+};
+
+export default courseImages;
